requirements.txt
python-telegram-bot==13.15
requests==2.31.0
feedparser==6.0.10
google-generativeai==0.3.2
beautifulsoup4==4.12.2
python-dotenv==1.0.0
APScheduler==3.6.3
pytz==2023.3 